[STEP1] 추가 수정 	- 폴더링 	- 파일분리 	- 사용하지 않는 메서드 제거
