+++ WeatherForecast/Foundation/WeatherForecaseInfo/Properties/MainInfo.swift
@@ -1,0 +1,14 @@
+//
+//  MainInfo.swift
+//  WeatherForecast
+//
+//  Created by 김창규 on 1/30/24.
+//
+
+import Foundation
+
+// MARK: - MainClass
+struct MainInfo: Decodable {
+    let temp, feelsLike, tempMin, tempMax: Double
+    let pressure, seaLevel, grndLevel, humidity, pop: Double
+}
