+++ WeatherForecast/Foundation/WeatherForecaseInfo/Properties/Weather.swift
@@ -1,0 +1,15 @@
+//
+//  WeatherForecast - Weather.swift
+//  Created by yagom.
+//  Copyright © yagom. All rights reserved.
+// 
+
+import Foundation
+
+// MARK: - Weather
+struct Weather: Decodable {
+    let id: Int
+    let main: String
+    let description: String
+    let icon: String
+}
