+++ WeatherForecast/WeatherForecast/WeatherList/Models/Protocols/WeatherInfoCoordinator.swift
@@ -8,13 +8,13 @@
 import Foundation
 
 protocol WeatherInfoCoordinator {
-    var tempUnit: TempUnit { get set }
+    var tempUnit: TempUnit { get }
     var weatherForecastInfo: [WeatherForecastInfo]? { get }
     func setWeatherJSON(json: WeatherJSON)
     func getCityInfo() -> City?
     func getWeatherForecastInfo(at index: Int) -> WeatherForecastInfo?
     func getTempUnit() -> TempUnit
     func getTemp(at index: Int) -> String?
-    func changeTempUnit()
+    func toggleTempUnit()
     func getTempExpressionTitle() -> String
 }
