+++ WeatherForecast/WeatherForecast/WeatherList/Models/Structs/City.swift
@@ -1,0 +1,18 @@
+//
+//  City.swift
+//  WeatherForecast
+//
+//  Created by 김창규 on 1/30/24.
+//
+
+import Foundation
+
+// MARK: - Ci
+struct City: Decodable {
+    let id: Int
+    let name: String
+    let coord: Coord
+    let country: String
+    let population, timezone: Int
+    let sunrise, sunset: TimeInterval
+}
