+++ WeatherForecast/WeatherForecast/WeatherList/Models/Structs/Coord.swift
@@ -1,0 +1,12 @@
+//
+//  Coord.swift
+//  WeatherForecast
+//
+//  Created by 김창규 on 1/30/24.
+//
+
+import Foundation
+// MARK: - Coord
+struct Coord: Decodable {
+    let lat, lon: Double
+}
